Update dorker.py
Added/updated the following in lines 87-90:
    subdomains = '*.' + targetUrl
    subsubdomains = '*.' + targetUrl
    awsFiles = targetUrl + '+intitle:"index.of" aws/'
    s3Buckets = targetUrl + '+intitle:"index.of" /aws.s3/'

Fixed S3Buckets to read s3Buckets in line 92.

Changed target to targetUrl in line 98. When target was being passed to the dork function it was being defined as targetUrl (line 65). This fixed the issue the script was having where it keep executing line 96-99.